Allow overriding the source field in log entries
The source field is hardcoded to include the word "heroku" but as
far as I can see there's nothing in the repo that requires apps
to be hosted on Heroku in order to use n-logger.

I'd like to integrate n-logger with a couple of apps that are
hosted on FT Platform. These apps currently log to the filesystem
and splunkd then collects the logs to send them to Splunk. Integrating
n-logger will make it much easier to move these apps to Heroku in the
future.

This commit will allow those apps to use n-logger while hosted on FT
Platform without making the source field confusing. 🐿 v2.12.5