Refactor not to use isomorphic fetch package
This refactors the package not to use the `isomorphic-fetch` package as this pollutes the global scope.

After this change a global fetch instance will still be used if available (such as apps which use n-express), otherwise it will fall back to its own `node-fetch` implementation.

I believe this is a non-breaking change as it is removing an undocumented side-effect. 🐿 v2.13.0